fix np.float error

+++ learning/custom_dataset.py
@@ -80,7 +80,7 @@ def preprocess_pointclouds(SEMA3D_PATH):
             if file.endswith(".h5"):
                 f = h5py.File(pathD + file, 'r')
                 xyz = f['xyz'][:]
-                rgb = f['rgb'][:].astype(np.float)
+                rgb = f['rgb'][:].astype(np.float32)
                 elpsv = np.stack([ f['xyz'][:,2][:], f['linearity'][:], f['planarity'][:], f['scattering'][:], f['verticality'][:] ], axis=1)
 
                 # rescale to [-0.5,0.5]; keep xyz
